// apps/api/src/sandbox/enums/job-type.enum.ts
/*
 * Copyright 2025 Daytona Platforms Inc.
 * SPDX-License-Identifier: AGPL-3.0
 */

export enum JobType {
  CREATE_SANDBOX = 'CREATE_SANDBOX',
  START_SANDBOX = 'START_SANDBOX',
  STOP_SANDBOX = 'STOP_SANDBOX',
  DESTROY_SANDBOX = 'DESTROY_SANDBOX',
  DESTROY_SANDBOX_WORKSPACE = 'DESTROY_SANDBOX_WORKSPACE',
  RECOVER_SANDBOX_WORKSPACE = 'RECOVER_SANDBOX_WORKSPACE',
  RESIZE_SANDBOX = 'RESIZE_SANDBOX',
  CREATE_BACKUP = 'CREATE_BACKUP',
  BUILD_SNAPSHOT = 'BUILD_SNAPSHOT',
  PULL_SNAPSHOT = 'PULL_SNAPSHOT',
  RECOVER_SANDBOX = 'RECOVER_SANDBOX',
  INSPECT_SNAPSHOT_IN_REGISTRY = 'INSPECT_SNAPSHOT_IN_REGISTRY',
  REMOVE_SNAPSHOT = 'REMOVE_SNAPSHOT',
  UPDATE_SANDBOX_NETWORK_SETTINGS = 'UPDATE_SANDBOX_NETWORK_SETTINGS',
  SNAPSHOT_SANDBOX = 'SNAPSHOT_SANDBOX',
  FORK_SANDBOX = 'FORK_SANDBOX',
  PAUSE_SANDBOX = 'PAUSE_SANDBOX',
}